# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend Developer Challenger API Docs.",
        "title": "Backend Developer Challenger API Docs",
        "contact": {
            "name": "Vinnicius Santos"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "paths": {
        "/api/v1/clients": {
            "get": {
                "description": "Busca todos os clientes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Busca todos os clientes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseshttp.RestSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseshttp.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responseshttp.RestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responseshttp.RestErr"
                        }
                    }
                }
            },
            "post": {
                "description": "Criar um novo cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "client"
                ],
                "summary": "Criar um novo cliente",
                "parameters": [
                    {
                        "description": "Client",
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateClientRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responseshttp.RestSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseshttp.RestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responseshttp.RestErr"
                        }
                    }
                }
            }
        },
        "/api/v1/clients/{id}": {
            "get": {
                "description": "Busca um cliente pelo id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clientes"
                ],
                "summary": "Busca um cliente pelo id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id do cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseshttp.RestSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseshttp.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responseshttp.RestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responseshttp.RestErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.CreateClientRequestDTO": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "health_problems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.CreateHealthProblemRequestDTO"
                    }
                },
                "name": {
                    "type": "string"
                },
                "sexo": {
                    "type": "string"
                }
            }
        },
        "request.CreateHealthProblemRequestDTO": {
            "type": "object",
            "properties": {
                "grau": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "responseshttp.Causes": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "responseshttp.RestErr": {
            "type": "object",
            "properties": {
                "causes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responseshttp.Causes"
                    }
                },
                "code": {
                    "type": "integer"
                },
                "err": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "responseshttp.RestSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}